requirements.txt
absl-py==0.15.0
atari-py==0.2.9
cffi==1.14.6
Cython==0.29.24
grpcio==1.37.1
gym==0.19.0
h5py==3.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
mpi4py==3.1.1
mujoco-py==2.0.2.13
numpy==1.19.5
opencv-python==4.5.4.58
pandas==1.1.5
pickleshare==0.7.5
protobuf==3.15.8
psutil==5.8.0
pyOpenSSL==21.0.0
scikit-learn==0.24.2
scipy==1.5.3
six==1.15.0
stable-baselines3==1.3.0
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
Theano==1.0.5
torch==1.10.0
tornado==6.1
tqdm==4.62.3
wheel==0.37.0
mkl
